Update Russian.lng
Russian translation v.XIS220827002Ru according present English version, but in the accordance with the context and traditions. Because English version has lot of problems with matching of tag's names and translations, for example with Zn industrial enterprises. Therefore, the Russian translation is based on the meaning of the English translation names, but not on the tag's names. Fixed mistakes. Added cases.